# api/swagger/swaggerDoc.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.4.0",
    "title": "dgx-patron-creator-service"
  },
  "host": "localhost:3001",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v0.1/patrons": {
      "x-swagger-router-controller": "patrons",
      "post": {
        "tags": [
          "patrons"
        ],
        "summary": "Create a Patron",
        "description": "Create a new patron account via NYPL's Simplified Card Creator.",
        "operationId": "patrons_creatorV01",
        "parameters": [
          {
            "name": "patron_data",
            "in": "body",
            "description": "The information of the new patron",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PatronsCreatorDataV01"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/PatronsCreatorResponseV01"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/400ErrorResponseV01"
            }
          },
          "500": {
            "description": "Generic server error",
            "schema": {
              "$ref": "#/definitions/500ErrorResponseV01"
            }
          }
        },
        "security": [
          {
            "api_auth": [
              "openid write:patron offline_access api"
            ]
          }
        ]
      }
    },
    "/v0.2/patrons": {
      "x-swagger-router-controller": "patrons",
      "post": {
        "tags": [
          "patrons"
        ],
        "summary": "Create a Patron",
        "description": "Create a new patron account via the ILS.",
        "operationId": "patrons_creatorV02",
        "parameters": [
          {
            "name": "patron_data",
            "in": "body",
            "description": "The information of the new patron",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PatronsCreatorDataV02"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/PatronsCreatorResponseV02"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/400ErrorResponseV02"
            }
          },
          "500": {
            "description": "Generic server error",
            "schema": {
              "$ref": "#/definitions/500ErrorResponseV02"
            }
          }
        },
        "security": [
          {
            "api_auth": [
              "openid write:patron offline_access api"
            ]
          }
        ]
      }
    },
    "/v0.3/patrons/dependents": {
      "x-swagger-router-controller": "patrons",
      "post": {
        "tags": [
          "patrons"
        ],
        "summary": "Create a dependent juvenile account",
        "description": "Create a dependent juvenile account based off a parent's patron account",
        "operationId": "patrons_dependentsV03",
        "parameters": [
          {
            "name": "barcode_and_input",
            "in": "body",
            "description": "The 14-digit barcode of the existing parent patron and the dependent's account information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PatronsDependentsDataV03"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/PatronsDependentsResponseV03"
            }
          },
          "400": {
            "description": "Invalid request",
            "schema": {
              "$ref": "#/definitions/400ErrorResponseV03"
            }
          },
          "502": {
            "description": "ILS Integration Error",
            "schema": {
              "$ref": "#/definitions/502ILSIntegrationErrorV03"
            }
          }
        },
        "security": [
          {
            "api_auth": [
              "openid write:patron offline_access api"
            ]
          }
        ]
      }
    },
...
    "/v0.1/validations/username": {
      "x-swagger-router-controller": "validations/username",
      "post": {
        "tags": [
          "validations"
        ],
        "summary": "Create a Patron username validation",
        "description": "Makes a call to NYPL's Simplified Card Creator API for patron username validation.",
        "operationId": "username",
        "parameters": [
          {
            "name": "username",
            "in": "body",
            "description": "The username to be validated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UsernameData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/UsernameResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/400ErrorResponse"
            }
          },
          "500": {
            "description": "Generic server error",
            "schema": {
              "$ref": "#/definitions/500ErrorResponse"
            }
          }
        },
        "security": [
          {
            "api_auth": [
              "openid write:patron offline_access api"
            ]
          }
        ]
      }
    },
    "/v0.1/validations/address": {
      "x-swagger-router-controller": "validations/address",
      "post": {
        "tags": [
          "validations"
        ],
        "summary": "Create a Patron address validation",
        "description": "Makes a call to NYPL's Simplified Card Creator API for patron address validation.",
        "operationId": "address",
        "parameters": [
          {
            "name": "address_data",
            "in": "body",
            "description": "The address to be validated",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddressData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/AddressResponse"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/400ErrorResponse"
            }
          },
          "500": {
            "description": "Generic server error",
            "schema": {
              "$ref": "#/definitions/500ErrorResponse"
            }
          }
        },
        "security": [
          {
            "api_auth": [
              "openid write:patron offline_access api"
            ]
          }
        ]
      }
    }
  },
  "definitions": {
    "PatronsCreatorDataV01": {
      "required": [
        "simplePatron"
      ],
      "properties": {
        "simplePatron": {
          "$ref": "#/definitions/SimplePatronV01"
        }
      }
    },
...
    "PatronsCreatorResponseV01": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/PatronsCreatorResponseModelV01"
        },
        "count": {
          "type": "number",
          "example": 1
        }
      }
    },
...
    "400ErrorResponseV01": {
      "properties": {
        "data": {
          "$ref": "#/definitions/400ErrorResponseContentV01"
        },
        "count": {
          "type": "number",
          "example": 0
        }
      }
    },
...
    "500ErrorResponseV01": {
      "properties": {
        "data": {
          "$ref": "#/definitions/500ErrorResponseContentV01"
        },
        "count": {
          "type": "number",
          "example": 0
        }
      }
    },
...
    "PatronsDependentsDataV03": {
      "required": [
        "barcode",
        "name",
        "username",
        "pin"
      ],
      "properties": {
        "barcode": {
          "type": "string",
          "example": "12345678912345"
        },
        "name": {
          "type": "string",
          "example": "Dependent First name, Last Name"
        },
        "username": {
          "type": "string",
          "example": "dependentUsername1"
        },
        "pin": {
          "type": "string",
          "example": "1234"
        }
      }
    },
    "PatronsCreatorDataV02": {
      "required": [
        "names",
        "patronType"
      ],
      "properties": {
        "names": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [ "TestLastName, TestFirstName" ]
        },
        "barcodes": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [ "barcode_2018_07_19_328pm" ]
        },
        "pin": {
          "type": "string",
          "example": "1234"
        },
        "expirationDate": {
          "type": "string",
          "example": "2019-01-01"
        },
        "birthDate": {
          "type": "string",
          "example": "1978-01-01"
        },
        "emails": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": [
            "test_email_2018_07_19_238_a@test.com",
            "test_email_2018_07_19_238_b@test.com"
          ]
        },
        "patronType": {
          "type": "number",
          "example": 151
        },
        "patronCodes": {
          "type": "object",
          "required": [
            "pcode1",
            "pcode2",
            "pcode3",
            "pcode4"
          ],
          "properties": {
            "pcode1": {
              "type": "string",
              "example": "s"
            },
            "pcode2": {
              "type": "string",
              "example": "f"
            },
            "pcode3": {
              "type": "number",
              "example": 5
            },
            "pcode4": {
              "type": "number",
              "example": 100
            }
          }
        },
        "blockInfo": {
          "type": "object",
          "required": [
            "code"
          ],
          "properties": {
            "code": {
              "type": "string",
              "example": "-"
            }
          }
        },
        "addresses": {
          "type": "array",
          "items": {
            "type": "object",
            "required": [
              "lines",
              "type"
            ],
            "properties": {
              "lines": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "ADDRESS LINE 1",
                  "ADDRESS LINE 2"
                ]
              },
              "type": {
                "type": "string",
                "example": "a"
              }
            }
          }
        },
        "phones": {
          "type": "array",
          "items": {
            "type": "object",
            "required": [
              "number",
              "type"
            ],
            "properties": {
              "number": {
                "type": "string",
                "example": "917-123-4567"
              },
              "type": {
                "type": "string",
                "example": "t"
              }
            }
          }
        }
      }
    },
    "PatronsCreatorResponseV02": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/PatronsCreatorResponseModelV02"
        },
        "count": {
          "type": "number",
          "example": 1
        }
      }
    },
...
    "400ErrorResponseV02": {
      "properties": {
        "status_code_from_ils": {
          "type": "number",
          "example": 400
        },
        "type": {
          "type": "string",
          "example": "invalid-request"
        },
        "message": {
          "type": "string",
          "example": "No data"
        },
        "detail": {
          "type": "object",
          "example": {
          }
        }
      }
    },
...
    "500ErrorResponseV02": {
      "properties": {
        "status_code_from_ils": {
          "type": "number",
          "example": 500
        },
        "type": {
          "type": "string",
          "example": "internal-server-error"
        },
        "message": {
          "type": "string",
          "example": "Internal server error"
        },
        "detail": {
          "type": "object",
          "example": {
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "api_auth": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://isso.nypl.org/oauth/authorize",
      "tokenUrl": "https://isso.nypl.org/oauth/token",
      "scopes": {
        "openid write:patron offline_access api": "Creating Patron and Validations access"
      }
    }
  }
}